+++ src/main/java/randommcsomethin/fallingleaves/init/Config.java
@@ -2,14 +2,11 @@ package randommcsomethin.fallingleaves.init;
 
 import me.sargunvohra.mcmods.autoconfig1u.AutoConfig;
 import me.sargunvohra.mcmods.autoconfig1u.serializer.GsonConfigSerializer;
-import org.apache.logging.log4j.Level;
-import org.apache.logging.log4j.core.config.Configurator;
+import randommcsomethin.fallingleaves.FallingLeavesClient;
 import randommcsomethin.fallingleaves.config.FallingLeavesConfig;
 import randommcsomethin.fallingleaves.config.LeafSettingsGuiProvider;
 
 import java.lang.reflect.Field;
-
-import static randommcsomethin.fallingleaves.LeafUtils.LOGGER;
 
 public class Config {
 
@@ -26,10 +23,7 @@ public class Config {
 
         CONFIG = FallingLeavesConfig.instance;
 
-        if (CONFIG.displayDebugData) {
-            if (LOGGER.getLevel().compareTo(Level.DEBUG) < 0) {
-                Configurator.setLevel(LOGGER.getName(), Level.DEBUG);
-            }
-        }
+        FallingLeavesClient.LOGGER.debug("Loaded configuration.");
+
     }
 }